refactor(sync): factor out class code
